move middleware to own package
